refactor: move tracing MethodHandles out of RocksDB into TracingSupport
RocksDB.java was accumulating MH_ fields for every feature area. Tracing's
six MHs plus their invoke/error-handling boilerplate now live in a new
package-private TracingSupport class that RocksDBTracingOperations calls
into directly, instead of routing through RocksDB.
