+++ CLAUDE.md
@@ -210,7 +210,7 @@ instance method listed below lives on the DB type (`ReadWriteDB`, `TtlDB`, …),
 | Backup Engine           | `BackupEngine.java`, `BackupEngineOptions.java`, `RestoreOptions.java`, `BackupInfo.java`, `BackupId.java`                |
 | Column Families         | `ColumnFamilyHandle.java`, `ColumnFamilyDescriptor.java`; `RocksDB.openReadWrite`, `listColumnFamilies`; CF overloads on `ReadWriteDB` and `WriteBatch`; CF overloads on `ReadOnlyDB`, `TtlDB`, `BlobDB`, `SecondaryDB`, `TransactionDB`, `OptimisticTransactionDB`; `Transaction` CF put/delete/get/getForUpdate/newIterator; multi-CF open for all DB types (including `RocksDB.openSecondary`) |
 | Perf Context            | `PerfContext.java`, `PerfLevel.java`, `PerfMetric.java`; thread-local; `setPerfLevel`, `reset`, `metric`, `report`       |
-| Tracing & Replay        | `TraceOptions.java`, `TraceFilter.java`; `RocksDBTracingOperations.startTrace`/`endTrace` — its own interface (not folded into `RocksDBWriteOperations`) since tracing captures reads too and works on read-only/secondary handles, implemented by the same set of types as `RocksDBMonitoringOperations` (`ReadWriteDB`/`TtlDB`/`BlobDB`/`OptimisticTransactionDB`/`ReadOnlyDB`/`SecondaryDB`/`TransactionDB`); `Replayer.java`, `ReplayOptions.java` — replays a captured trace against a target database. Same interface also has `startIoTrace`/`endIoTrace` (file I/O, reuses `TraceOptions`) and `startBlockCacheTrace`/`endBlockCacheTrace` (block cache accesses; `TraceOptions` overload or `BlockCacheTraceOptions.java`/`BlockCacheTraceWriterOptions.java` overload, mirroring `rocksdb_start_block_cache_trace`/`_with_options`) |
+| Tracing & Replay        | `TraceOptions.java`, `TraceFilter.java`; `RocksDBTracingOperations.startTrace`/`endTrace` — its own interface (not folded into `RocksDBWriteOperations`) since tracing captures reads too and works on read-only/secondary handles, implemented by the same set of types as `RocksDBMonitoringOperations` (`ReadWriteDB`/`TtlDB`/`BlobDB`/`OptimisticTransactionDB`/`ReadOnlyDB`/`SecondaryDB`/`TransactionDB`); `Replayer.java`, `ReplayOptions.java` — replays a captured trace against a target database. Same interface also has `startIoTrace`/`endIoTrace` (file I/O, reuses `TraceOptions`) and `startBlockCacheTrace`/`endBlockCacheTrace` (block cache accesses; `TraceOptions` overload or `BlockCacheTraceOptions.java`/`BlockCacheTraceWriterOptions.java` overload, mirroring `rocksdb_start_block_cache_trace`/`_with_options`). All six default methods delegate to package-private statics on `TracingSupport.java`, not `RocksDB` — the tracing `MH_` fields and error-handling boilerplate live there instead of growing `RocksDB.java` further; an interface can't hold `private static final` fields itself, so this sibling class is the closest equivalent |
 | Event listeners         | `EventNotifier.java` (8 no-op-default callbacks), `EventNotifierBridge.java` (FFM plumbing, one global upcall stub per callback + `UpcallRegistry` id-through-`state`); payload views `FlushJobInfo.java`, `CompactionJobInfo.java`, `ExternalFileIngestionInfo.java`, `MemTableInfo.java`, `WriteStallInfo.java`; enums `FlushReason.java`, `CompactionReason.java`, `BackgroundErrorReason.java`, `WriteStallCondition.java`; wired via `Options.addEventListener`. The two subcompaction callbacks are deliberately not exposed, but still get a shared no-op stub — `db/c.cc` calls them without a null check |
 | Upcall thread lifetime  | `BackgroundUpcallThreads.java` — shutdown hook that drains RocksDB's background threads so `System.exit()` cannot deadlock once an upcall has attached one to the JVM. `installShutdownDrain()` from every callback-registration site (`EventNotifierBridge.attach`, `Logger.newCallbackLogger`, `MergeOperator.Custom.create`), `track()` from every upcall dispatch that RocksDB can invoke on a background thread. Any new callback-based feature must call both |
 | Command-line tools      | `NativeTool.java` (core — subprocess extraction/execution infra, not FFM; nests the `Result` record); `tools/ldb` (`rocksdbffm-tool-ldb`: `LdbTool.java`, `Main.java`) and `tools/sst-dump` (`rocksdbffm-tool-sstdump`: `SstDumpTool.java`, `SstDumpCommand.java`, `Main.java`) wrap the bundled `ldb`/`sst_dump` binaries — not in `c.h`, no FFM path exists for them. Each tool module is a normal jar (native deps at test scope only, mirroring `core`) with both a programmatic API and a `Main-Class` for `java -jar`/`java -cp ... Main` passthrough. On macOS/Linux, `scripts/build-rocksdb.sh` builds them `LIB_MODE=shared` (~300KB/binary + one shared `librocksdb_tools.*`, vs ~15MB/binary static) with `USE_RTTI=1` (required by `ldb_cmd.cc`'s `Customizable`-based parsing; without it, DEBUG_LEVEL=0's default `-fno-rtti` produces `.o`s with vtables but no typeinfo, a hard link error — must apply from the `shared_lib` step onward since object files are reused across invocations) and `-Wl,-Bsymbolic` on both the `shared_lib` and `ldb`/`sst_dump` link steps, required to avoid a heap-corrupting libc++ symbol-interposition crash — see [ADR 0008](docs/adr/0008-ldb-sst-dump-dynamic-linking.md). On `windows-*`, `scripts/build-rocksdb-windows.sh` builds `ldb.exe`/`sst_dump.exe` via CMake's `WITH_CORE_TOOLS`/`USE_RTTI` options instead — statically linked (RocksDB's CMake only selects the shared lib as the tools' link target off Windows, a vendored-source constraint not worked around), so noticeably larger than the POSIX binaries but with no runtime library dependency at all |
